Project/fyp/requirements.txt
django==5.0.4
asgiref==3.8.1
autopep8==2.1.0
toml==0.10.2
whitenoise==6.6.0
django-environ==0.11.2
fontawesomefree==6.5.1
django-notifications-hq==1.8.3
mysqlclient==2.2.4
channels==4.1.0
django-debug-toolbar==4.3.0
kolo